Working emergency info controllers

+++ lib/litelist_web/controllers/emergency_information_controller.ex
@@ -2,20 +2,34 @@ defmodule LitelistWeb.EmergencyInformationController do
   use LitelistWeb, :controller
 
   alias Litelist.Posts
-  alias Litelist.Posts.EmergencyInformation
+  alias Litelist.Posts.Post
+
+  alias LitelistWeb.Utils.SharedUtils
+
+  @post_type "emergency_information"
+  @permitted_params ["contact_info",
+    "description",
+    "title",
+    "url",
+    "organization_name"
+  ]
 
   def index(conn, _params) do
-    emergency_information = Posts.list_emergency_information()
+    emergency_information = Posts.list_posts_by_type(@post_type)
     render(conn, "index.html", emergency_information: emergency_information)
   end
 
   def new(conn, _params) do
-    changeset = Posts.change_emergency_information(%EmergencyInformation{})
+    changeset = Posts.change_post(%Post{})
     render(conn, "new.html", changeset: changeset)
   end
 
-  def create(conn, %{"emergency_information" => emergency_information_params}) do
-    case Posts.create_emergency_information(emergency_information_params) do
+  def create(conn, %{"post" => emergency_information_params}) do
+    emergency_information_params = emergency_information_params
+      |> SharedUtils.permitted_params(@permitted_params)
+      |> SharedUtils.add_generated_params(conn, @post_type, :create)
+
+    case Posts.create_post(emergency_information_params) do
       {:ok, emergency_information} ->
         conn
         |> put_flash(:info, "Emergency information created successfully.")
@@ -26,35 +40,60 @@ defmodule LitelistWeb.EmergencyInformationController do
   end
 
   def show(conn, %{"id" => id}) do
-    emergency_information = Posts.get_emergency_information!(id)
-    render(conn, "show.html", emergency_information: emergency_information)
+    emergency_information = Posts.get_post!(id)
+    if SharedUtils.match_type?(emergency_information, @post_type) do
+      render(conn, "show.html", emergency_information: emergency_information)
+    else
+      unauthorized_redirect(conn)
+    end
   end
 
   def edit(conn, %{"id" => id}) do
-    emergency_information = Posts.get_emergency_information!(id)
-    changeset = Posts.change_emergency_information(emergency_information)
-    render(conn, "edit.html", emergency_information: emergency_information, changeset: changeset)
+    emergency_information = Posts.get_post!(id)
+    if SharedUtils.permission?(conn.assigns.current_neighbor, emergency_information, @post_type) do
+      changeset = Posts.change_post(emergency_information)
+      render(conn, "edit.html", emergency_information: emergency_information, changeset: changeset)
+    else
+      unauthorized_redirect(conn)
+    end
   end
 
-  def update(conn, %{"id" => id, "emergency_information" => emergency_information_params}) do
-    emergency_information = Posts.get_emergency_information!(id)
-
-    case Posts.update_emergency_information(emergency_information, emergency_information_params) do
-      {:ok, emergency_information} ->
-        conn
-        |> put_flash(:info, "Emergency information updated successfully.")
-        |> redirect(to: emergency_information_path(conn, :show, emergency_information))
-      {:error, %Ecto.Changeset{} = changeset} ->
-        render(conn, "edit.html", emergency_information: emergency_information, changeset: changeset)
+  def update(conn, %{"id" => id, "post" => emergency_information_params}) do
+    emergency_information = Posts.get_post!(id)
+    if SharedUtils.permission?(conn.assigns.current_neighbor, emergency_information, @post_type) do
+      emergency_information_params = emergency_information_params
+        |> SharedUtils.permitted_params(@permitted_params)
+        |> SharedUtils.add_generated_params(:update)
+      
+      case Posts.update_post(emergency_information, emergency_information_params) do
+        {:ok, emergency_information} ->
+          conn
+          |> put_flash(:info, "Emergency information updated successfully.")
+          |> redirect(to: emergency_information_path(conn, :show, emergency_information))
+        {:error, %Ecto.Changeset{} = changeset} ->
+          render(conn, "edit.html", emergency_information: emergency_information, changeset: changeset)
+      end
+    else
+      unauthorized_redirect(conn)
     end
   end
 
   def delete(conn, %{"id" => id}) do
-    emergency_information = Posts.get_emergency_information!(id)
-    {:ok, _emergency_information} = Posts.delete_emergency_information(emergency_information)
+    emergency_information = Posts.get_post!(id)
+    if SharedUtils.permission?(conn.assigns.current_neighbor, emergency_information, @post_type) do
+      {:ok, _emergency_information} = Posts.delete_emergency_information(emergency_information)
 
+      conn
+      |> put_flash(:info, "Emergency information deleted successfully.")
+      |> redirect(to: emergency_information_path(conn, :index))
+    else
+      unauthorized_redirect(conn)
+    end
+  end
+
+  defp unauthorized_redirect(conn) do
     conn
-    |> put_flash(:info, "Emergency information deleted successfully.")
+    |> put_flash(:error, "Unauthorized.")
     |> redirect(to: emergency_information_path(conn, :index))
   end
 end
